requirements.txt
anthropic>=0.30.0
fal-client>=0.5.0
playwright>=1.40.0
Pillow>=10.0.0
colorthief>=0.2.1
requests>=2.31.0
moviepy>=1.0.3
python-dotenv>=1.0.0
